Changed error handling
* Gracefully handle a failure to get an HTTP resource
* Use `rand::choose()` instead of using slice indices

Note that choose() can still fail if the slice it's called on is empty. Because the vector they're called on is changed
only at compile time, both are safe to unwrap if the application is tested to pick each of the options. Typically using
the direct index in a vector or slice is considered a no-no (anit-pattern) in Rust.